+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Spring 4.3.10
     * Jackson 2.8.10
 * Create example implementations in extra repo
+* Support EC2 Parameter Store as Config Source (#54)
 
 # Version 1.22
 * enhanced support to get versioned element from MongoDB DAOs